requirements.txt
geopy==2.0.0
aiogram==2.10.1
aiohttp==3.7.3
aioredis==1.3.1
python-dotenv==0.15.0
asyncio==3.4.3
asyncpg==0.21.0
gino
sqlalchemy
